requirements.txt
openai
python-dotenv
pyyaml
rich
asyncio